[Refactor] Restructure and clean the majority of codebase

+++ height_mapping_core/src/FastHeightFilter.cpp
@@ -2,13 +2,12 @@
 
 namespace height_mapping {
 
-FastHeightFilter::FastHeightFilter(float min_height, float max_height)
+FastHeightFilter::FastHeightFilter(double min_height, double max_height)
     : minZ_(min_height), maxZ_(max_height) {}
 
 template <typename PointT>
-void FastHeightFilter::filter(
-    const typename pcl::PointCloud<PointT>::Ptr &input,
-    typename pcl::PointCloud<PointT>::Ptr &output) {
+void FastHeightFilter::filter(const typename pcl::PointCloud<PointT>::Ptr &input,
+                              typename pcl::PointCloud<PointT>::Ptr &output) {
   output->clear();
   output->reserve(input->size());
 
@@ -21,12 +20,10 @@ void FastHeightFilter::filter(
 }
 
 // Explicit instantiation for the types we use
-template void
-FastHeightFilter::filter<Laser>(const pcl::PointCloud<Laser>::Ptr &input,
-                                pcl::PointCloud<Laser>::Ptr &output);
+template void FastHeightFilter::filter<Laser>(const pcl::PointCloud<Laser>::Ptr &input,
+                                              pcl::PointCloud<Laser>::Ptr &output);
 
-template void
-FastHeightFilter::filter<Color>(const pcl::PointCloud<Color>::Ptr &input,
-                                pcl::PointCloud<Color>::Ptr &output);
+template void FastHeightFilter::filter<Color>(const pcl::PointCloud<Color>::Ptr &input,
+                                              pcl::PointCloud<Color>::Ptr &output);
 
 } // namespace height_mapping
